모달 UI 개선 및 레이아웃 수정
