+++ src/pages/user_event/EventDetail.tsx
