+++ src/pages/user_event/Reviews.tsx
@@ -129,9 +129,8 @@ export const Reviews = () => {
     return Array.from({ length: 5 }, (_, index) => (
       <span
         key={index}
-        className={`text-base leading-6 ${
-          index < rating ? "text-[#ffd700]" : "text-[#dddddd]"
-        }`}
+        className={`text-base leading-6 ${index < rating ? "text-[#ffd700]" : "text-[#dddddd]"
+          }`}
       >
         ★
       </span>
@@ -143,10 +142,10 @@ export const Reviews = () => {
       prevReviews.map(review =>
         review.id === reviewId
           ? {
-              ...review,
-              isLiked: !review.isLiked,
-              likeCount: review.isLiked ? review.likeCount - 1 : review.likeCount + 1
-            }
+            ...review,
+            isLiked: !review.isLiked,
+            likeCount: review.isLiked ? review.likeCount - 1 : review.likeCount + 1
+          }
           : review
       )
     );
@@ -162,15 +161,15 @@ export const Reviews = () => {
       alert("신고 사유를 입력해주세요.");
       return;
     }
-    
+
     // 실제로는 API 호출하여 신고 데이터를 서버에 전송
     console.log(`관람평 ID ${selectedReviewId} 신고: ${reportReason}`);
-    
+
     // 모달 닫기 및 상태 초기화
     setShowReportModal(false);
     setSelectedReviewId(null);
     setReportReason("");
-    
+
     // 신고 접수 완료 메시지
     alert("신고가 접수되었습니다.");
   };
@@ -189,7 +188,7 @@ export const Reviews = () => {
   const calculateAverageRating = (): string => {
     const visibleReviews = reviews.filter(review => !review.isHidden);
     if (visibleReviews.length === 0) return "0.00";
-    
+
     const totalRating = visibleReviews.reduce((sum, review) => sum + review.rating, 0);
     return (totalRating / visibleReviews.length).toFixed(2);
   };
@@ -202,39 +201,38 @@ export const Reviews = () => {
         <h3 className="text-2xl font-semibold text-[#212121]">
           관람평
         </h3>
-                 <div className="flex items-center gap-2">
-           <div className="flex gap-1">
-             {Array.from({ length: 5 }, (_, index) => {
-               const rating = parseFloat(averageRating);
-               const fullStars = Math.floor(rating);
-               const hasPartialStar = index === fullStars && rating % 1 > 0;
-               const isFullStar = index < fullStars;
-               const isPartialStar = hasPartialStar;
-               
-               return (
-                 <span
-                   key={index}
-                   className={`text-xl leading-6 relative ${
-                     isFullStar ? "text-[#ffd700]" : "text-[#dddddd]"
-                   }`}
-                 >
-                   ★
-                   {isPartialStar && (
-                     <span 
-                       className="absolute top-0 left-0 text-[#ffd700] overflow-hidden"
-                       style={{ width: `${(rating % 1) * 100}%` }}
-                     >
-                       ★
-                     </span>
-                   )}
-                 </span>
-               );
-             })}
-           </div>
-           <span className="text-lg font-medium text-[#212121]">
-             {averageRating}
-           </span>
-         </div>
+        <div className="flex items-center gap-2">
+          <div className="flex gap-1">
+            {Array.from({ length: 5 }, (_, index) => {
+              const rating = parseFloat(averageRating);
+              const fullStars = Math.floor(rating);
+              const hasPartialStar = index === fullStars && rating % 1 > 0;
+              const isFullStar = index < fullStars;
+              const isPartialStar = hasPartialStar;
+
+              return (
+                <span
+                  key={index}
+                  className={`text-xl leading-6 relative ${isFullStar ? "text-[#ffd700]" : "text-[#dddddd]"
+                    }`}
+                >
+                  ★
+                  {isPartialStar && (
+                    <span
+                      className="absolute top-0 left-0 text-[#ffd700] overflow-hidden"
+                      style={{ width: `${(rating % 1) * 100}%` }}
+                    >
+                      ★
+                    </span>
+                  )}
+                </span>
+              );
+            })}
+          </div>
+          <span className="text-lg font-medium text-[#212121]">
+            {averageRating}
+          </span>
+        </div>
       </div>
 
       <div className="space-y-4">
@@ -273,11 +271,10 @@ export const Reviews = () => {
               <div className="flex items-center gap-4">
                 <button
                   onClick={() => handleLike(review.id)}
-                  className={`flex items-center gap-2 text-sm font-normal transition-colors ${
-                    review.isLiked 
-                      ? "text-red-500" 
-                      : "text-[#00000099] hover:text-red-500"
-                  }`}
+                  className={`flex items-center gap-2 text-sm font-normal transition-colors ${review.isLiked
+                    ? "text-red-500"
+                    : "text-[#00000099] hover:text-red-500"
+                    }`}
                 >
                   <span className="text-lg">
                     {review.isLiked ? "❤️" : "🤍"}
@@ -304,37 +301,34 @@ export const Reviews = () => {
             <button
               onClick={() => handlePageChange(currentPage - 1)}
               disabled={currentPage === 1}
-              className={`px-3 py-2 rounded border text-sm font-normal transition-colors ${
-                currentPage === 1
-                  ? "text-gray-400 border-gray-200 cursor-not-allowed"
-                  : "text-[#00000099] border-[#00000033] hover:bg-gray-50"
-              }`}
+              className={`px-3 py-2 rounded border text-sm font-normal transition-colors ${currentPage === 1
+                ? "text-gray-400 border-gray-200 cursor-not-allowed"
+                : "text-[#00000099] border-[#00000033] hover:bg-gray-50"
+                }`}
             >
               &lt;
             </button>
-            
+
             {Array.from({ length: totalPages }, (_, index) => index + 1).map((page) => (
               <button
                 key={page}
                 onClick={() => handlePageChange(page)}
-                className={`px-3 py-2 rounded border text-sm font-normal transition-colors ${
-                  currentPage === page
-                    ? "bg-black text-white border-black"
-                    : "text-[#00000099] border-[#00000033] hover:bg-gray-50"
-                }`}
+                className={`px-3 py-2 rounded border text-sm font-normal transition-colors ${currentPage === page
+                  ? "bg-black text-white border-black"
+                  : "text-[#00000099] border-[#00000033] hover:bg-gray-50"
+                  }`}
               >
                 {page}
               </button>
             ))}
-            
+
             <button
               onClick={() => handlePageChange(currentPage + 1)}
               disabled={currentPage === totalPages}
-              className={`px-3 py-2 rounded border text-sm font-normal transition-colors ${
-                currentPage === totalPages
-                  ? "text-gray-400 border-gray-200 cursor-not-allowed"
-                  : "text-[#00000099] border-[#00000033] hover:bg-gray-50"
-              }`}
+              className={`px-3 py-2 rounded border text-sm font-normal transition-colors ${currentPage === totalPages
+                ? "text-gray-400 border-gray-200 cursor-not-allowed"
+                : "text-[#00000099] border-[#00000033] hover:bg-gray-50"
+                }`}
             >
               &gt;
             </button>
@@ -359,50 +353,50 @@ export const Reviews = () => {
           <p className="text-sm text-black font-normal">
             • 스포일러가 포함된 내용은 다른 관람객을 위해 주의해주세요.
           </p>
-                 </div>
-       </div>
+        </div>
+      </div>
 
-       {/* 신고 모달 */}
-       {showReportModal && (
-         <div className="fixed inset-0 flex items-center justify-center z-50">
-           <div className="bg-white w-[411px] rounded-[10px] shadow-lg border border-gray-200">
-             <div className="p-8">
-               <div className="text-center mb-6">
-                 <h3 className="[font-family:'Segoe_UI-Bold',Helvetica] font-bold text-black text-xl tracking-[0] leading-[30px] mb-4">
-                   신고하기
-                 </h3>
-                 <p className="[font-family:'Roboto-Regular',Helvetica] font-normal text-gray-600 text-base tracking-[0] leading-6 mb-6">
-                   신고 사유를 입력해주세요
-                 </p>
-               </div>
-               
-               <div className="mb-6">
-                 <textarea
-                   value={reportReason}
-                   onChange={(e) => setReportReason(e.target.value)}
-                   placeholder="신고 사유를 입력해주세요..."
-                   className="w-full p-4 border border-gray-300 rounded-lg resize-none h-32 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent [font-family:'Roboto-Regular',Helvetica] font-normal text-black text-base tracking-[0] leading-6"
-                 />
-               </div>
-               
-               <div className="flex gap-3">
-                 <button
-                   onClick={handleCloseModal}
-                   className="flex-1 px-4 py-3 border border-gray-300 rounded-lg text-gray-700 hover:bg-gray-50 transition-colors [font-family:'Roboto-SemiBold',Helvetica] font-semibold text-base tracking-[0] leading-6"
-                 >
-                   취소
-                 </button>
-                 <button
-                   onClick={handleSubmitReport}
-                   className="flex-1 px-4 py-3 bg-red-500 text-white rounded-lg hover:bg-red-600 transition-colors [font-family:'Roboto-SemiBold',Helvetica] font-semibold text-base tracking-[0] leading-6"
-                 >
-                   신고하기
-                 </button>
-               </div>
-             </div>
-           </div>
-         </div>
-       )}
-     </div>
-   );
- }; 
+      {/* 신고 모달 */}
+      {showReportModal && (
+        <div className="fixed inset-0 flex items-center justify-center z-[1001] bg-black bg-opacity-30">
+          <div className="bg-white w-[411px] rounded-[10px] shadow-lg border border-gray-200">
+            <div className="p-8">
+              <div className="text-center mb-6">
+                <h3 className="[font-family:'Segoe_UI-Bold',Helvetica] font-bold text-black text-xl tracking-[0] leading-[30px] mb-4">
+                  신고하기
+                </h3>
+                <p className="[font-family:'Roboto-Regular',Helvetica] font-normal text-gray-600 text-base tracking-[0] leading-6 mb-6">
+                  신고 사유를 입력해주세요
+                </p>
+              </div>
+
+              <div className="mb-6">
+                <textarea
+                  value={reportReason}
+                  onChange={(e) => setReportReason(e.target.value)}
+                  placeholder="신고 사유를 입력해주세요..."
+                  className="w-full p-4 border border-gray-300 rounded-lg resize-none h-32 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent [font-family:'Roboto-Regular',Helvetica] font-normal text-black text-base tracking-[0] leading-6"
+                />
+              </div>
+
+              <div className="flex gap-3">
+                <button
+                  onClick={handleCloseModal}
+                  className="flex-1 px-4 py-3 border border-gray-300 rounded-lg text-gray-700 hover:bg-gray-50 transition-colors [font-family:'Roboto-SemiBold',Helvetica] font-semibold text-base tracking-[0] leading-6"
+                >
+                  취소
+                </button>
+                <button
+                  onClick={handleSubmitReport}
+                  className="flex-1 px-4 py-3 bg-red-500 text-white rounded-lg hover:bg-red-600 transition-colors [font-family:'Roboto-SemiBold',Helvetica] font-semibold text-base tracking-[0] leading-6"
+                >
+                  신고하기
+                </button>
+              </div>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
