+++ src/pages/user_mypage/Favorites.tsx
