+++ src/pages/user_mypage/MyTickets.tsx
